Feedback CSR

+++ src/main/java/com/training360/cafebabeswebshop/feedback/Feedback.java
@@ -7,14 +7,14 @@ import java.time.LocalDateTime;
 
 public class Feedback {
 
-    private int id;
+    private long id;
     private LocalDateTime feedbackDate;
     private String feedback;
     private int rating;
     private User user;
     private Product product;
 
-    public Feedback(int id, String feedback, int rating, User user, Product product) {
+    public Feedback(long id, String feedback, int rating, User user, Product product) {
         this.id = id;
         this.feedbackDate = LocalDateTime.now();
         this.feedback = feedback;
@@ -23,7 +23,7 @@ public class Feedback {
         this.product = product;
     }
 
-    public Feedback(int id, LocalDateTime feedbackDate, String feedback, int rating, User user, Product product) {
+    public Feedback(long id, LocalDateTime feedbackDate, String feedback, int rating, User user, Product product) {
         this.id = id;
         this.feedbackDate = feedbackDate;
         this.feedback = feedback;
@@ -36,7 +36,7 @@ public class Feedback {
         return feedbackDate;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
